fix: define non authenticated content index home page
