--- apps/bytebank/src/app/home/page.module.scss
@@ -1,104 +1,0 @@
-.homePage {
-  display: flex;
-  flex-direction: column;
-  width: 100vw;
-  min-height: 100vh;
-  text-align: center;
-  overflow-x: hidden; 
-  background: linear-gradient(180deg, rgba(0, 77, 97, 1) 0%, rgba(255, 255, 255, 1) 100%);
-}
-
-.mobileButtonsWrapper {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  gap: 1rem;
-  margin-left: 15px;
-  margin-right: 15px;
-  margin-bottom: 32px;
-}
-
-.valueProposition {
-  margin-top: 42px;
-  width: 300px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 1rem;
-  text-align: center;
-}
-
-@media (min-width: 768px) and (max-width: 819px) {
-  .valueProposition {
-    width: 'calc(50% - 1rem)',
-  }
-}
-
-@media (min-width: 1024px) and (max-width: 1366px) {
-  .valueProposition {
-    margin-left: 60px;
-    margin-right: 60px;
-    width: calc(33.33% - 1rem);
-    max-width: calc(33.33% - 1rem);
-  }
-}
-
-.text {
-  color: var(--mui-palette-grey-600);
-  white-space: normal;
-  line-height: 1.5;
-  display: block;
-  word-wrap: break-word;
-  margin-bottom: 24px;
-  width: 100%;
-  max-width: 270px;
-  margin-left: 15px;
-  margin-right: 15px;
-
-  @media (min-width: 1366px) {
-    align-self: flex-end; 
-  }
-}
-
-.bannerText {
-  @extend .text;
-  color: var(--mui-palette-color-black); 
-  width: 100%;
-  max-width: 434px;
-}
-
-.banner {
-  margin-top: 24px;
-  margin-bottom: 42px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  @media (min-width: 1366px) {
-    flex-direction: row;
-    justify-content: center;
-    text-align: left;
-  }
-}
-
-.illustration {
-  width: 90vw;
-  max-width: 600px;
-  margin-top: 24px;
-  margin-left: 15px;
-  margin-right: 15px;
-
-  img {
-    width: 100%;
-    height: auto;
-    display: block;
-  }
-
-  @media (min-width: 1024px) and (max-width: 1366px) {
-    width: 661px;
-    margin-left: 60px;
-    margin-right: 60px;
-  }
-}
